requirements.txt
replicate==0.22.0
requests==2.31.0
python-dotenv==1.0.0
gradio==4.19.2
moviepy==1.0.3
opencv-python==4.9.0.80
elevenlabs==0.2.27
google-generativeai==0.3.2
anthropic==0.18.1
google-cloud-storage==2.14.0
Pillow==10.2.0 